requirements.txt
gradio-client
flask
flask_cors
python-dotenv